split utils and rename
